Create triangles.py
What is the probability of picking three "random" points that result in an obtuse (vs. acute) triangle?  This was a famous problem that showed, among other things, the need to specify what exactly we mean by "random".  I test this out with a uniform distribution and a normal distribution, where all six values are i.i.d. random variables (three for each axis).  

Note that I was a little lazy in recording this code and didn't include all of the necessary imports: I used ipython --pylab, which load in a whole mess of libraries, notably matplotlib.pyplot and numpy.  It shouldn't be too hard to figure out...