Create dadcircles2.py
My dad paints for a living.  Nowadays, mostly landscapes, but in the past he has done a lot of work with geometric paintings; one series was based on concentric circles.  This was a bit of a collabo with him, I wrote this script to generate patterns and automatically color them.