Create gf3.py
A few graphing functions.  The primary one creates pages of bar graphs.  A lot of the work was done to space the graphs correctly with respect to text.  More generally, it is for the sake of automating pretty mundane graphing so I can whip out lots of graphs easily, and yet have sufficient ability to customize them.